[feat] #17 전체적인 코드 refac
