+++ src/main/java/sopt/web1/activity/controller/ActivityController.java
@@ -5,14 +5,11 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import sopt.web1.activity.controller.dto.BestDto;
 import sopt.web1.activity.service.ActivityService;
 
-import java.util.List;
 import org.springframework.web.bind.annotation.PathVariable;
-import sopt.web1.activity.dto.ActResDto;
+import sopt.web1.activity.controller.dto.ActResDto;
 
 
 @RestController
@@ -22,13 +19,6 @@ import sopt.web1.activity.dto.ActResDto;
 public class ActivityController {
 
     private final ActivityService activityService;
-
-    @GetMapping("/best")
-    public ResponseEntity<BestDto> best(
-            @RequestParam final String category
-    ){
-        return ResponseEntity.ok(activityService.getBest(category));
-    }
 
     @GetMapping("/pick/activity/{activityId}")
     public ResponseEntity<ActResDto> getDetails(
