+++ src/main/java/sopt/web1/common/advice/ResponseWrappingAdvice.java
@@ -9,7 +9,7 @@ import org.springframework.http.server.ServerHttpRequest;
 import org.springframework.http.server.ServerHttpResponse;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseBodyAdvice;
-import sopt.web1.code.SuccessCode;
+import sopt.web1.common.code.SuccessCode;
 import sopt.web1.common.response.ResponseDto;
 
 @RestControllerAdvice
