+++ src/main/java/sopt/web1/exception/NotFoundCategoryException.java
@@ -1,7 +1,7 @@
 package sopt.web1.exception;
 
 import lombok.RequiredArgsConstructor;
-import sopt.web1.code.ErrorCode;
+import sopt.web1.common.code.ErrorCode;
 
 @RequiredArgsConstructor
 public class NotFoundCategoryException extends RuntimeException{
