+++ src/main/java/sopt/web1/home/controller/HomeController.java
@@ -1,4 +1,4 @@
-package sopt.web1.home;
+package sopt.web1.home.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -6,6 +6,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import sopt.web1.activity.controller.dto.BestResDto;
+import sopt.web1.home.controller.dto.PickResDto;
+import sopt.web1.home.servcie.HomeService;
 
 @RestController
 @RequestMapping("/api")
@@ -19,5 +22,12 @@ public class HomeController {
             @RequestParam final int bundleId
     ){
         return ResponseEntity.ok(homeService.getBundle(bundleId));
+    }
+
+    @GetMapping("/best")
+    public ResponseEntity<BestResDto> best(
+            @RequestParam final String category
+    ){
+        return ResponseEntity.ok(homeService.getBest(category));
     }
 }
